created GUIDisplay and guiFunctions.
seperated out methods from UMLBuilderController into the new created classes
