+++ src/main/java/umleditor/controller/MethodRenameController.java
@@ -2,7 +2,6 @@ package umleditor.controller;
 
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
-import javafx.scene.Node;
 import javafx.scene.control.TextArea;
 import javafx.scene.control.TextField;
 import umleditor.controller.utilities.Functions;
