+++ src/main/java/umleditor/controller/RelationshipArrowController.java
@@ -4,15 +4,13 @@ import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.control.TextArea;
 import javafx.scene.layout.AnchorPane;
-import umleditor.controller.utilities.Functions;
-import umleditor.model.utilities.Storage;
 import umleditor.view.gui.UMLNode;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class RelationshipArrowController {
-    List<umleditor.controller.MovableLine> lines = new ArrayList<>();
+    List<MovableLine> lines = new ArrayList<>();
     List<UMLNode> nodes = new ArrayList<>();
 
     @FXML
@@ -29,7 +27,7 @@ public class RelationshipArrowController {
     @FXML
     void addLine(ActionEvent event) {
 
-        MovableLine newLine = new umleditor.controller.MovableLine(anchorPane);
+        MovableLine newLine = new MovableLine(anchorPane);
         lines.add(newLine);
